split raiden.network.rpc.client
In respect to the proxies:

- Proxies used to expose a python interface that mimics the smart
contract interfaces are grouped under the proxies module
- Each proxy has its own module

In repect to the client:

- The magic is moved to smartcontract_proxy module
