+++ raiden/network/proxies/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+from raiden.network.proxies.channel_manager import ChannelManager  # NOQA
+from raiden.network.proxies.discovery import Discovery  # NOQA
+from raiden.network.proxies.netting_channel import NettingChannel  # NOQA
+from raiden.network.proxies.registry import Registry  # NOQA
+from raiden.network.proxies.token import Token  # NOQA
