+++ tools/deploy.py
@@ -9,9 +9,7 @@ from ethereum import slogging
 from ethereum._solidity import compile_contract
 from ethereum.utils import decode_hex
 
-from raiden.network.rpc.client import (
-    JSONRPCClient,
-)
+from raiden.network.rpc.client import JSONRPCClient
 from raiden.settings import GAS_PRICE
 from raiden.utils import get_contract_path
